Add AsyncLocalStorage driver (#39)
* Remove `bindEmitter` call

After doing research on what `bindEmitter` does, I could not see the
benefits of using it. At the time the hooks run, the transaction is already
over. There seem to be no reason to preserve store context for userland
use.

All current tests pass. There is no mention of preserved context in the
README, therefore this should be treated as internal change.

This change is motivated by the fact AsyncLocalStorage does not support
`bindEmitter` out of the box and without understanding the end goal it
is not feasible to properly reimplement it.

If at any point in the future a viable use-case is provided, the bindEmitter
can easily be restored.

* Add StorageLayer interface and implementations

This change introduces a concept of StorageLayer, an interface which
can be implemented using various storage providers. Currently the
implementations include AsyncLocalStorage and cls-hooked.

* Abstract direct cls-hooked use with StorageLayer

This commit abstracts the use of cls-hooked to StorageLayer.

It also updates the StorageLayer provider to always use cls-hooked
implementation, as AsyncLocalStorage implementation
does not pass tests yet.

* Implement AsyncLocalStorage

The change implements AsyncLocalStorage driver, which passes
all existing tests. As the actual logic relies on cls-hooked behavior
a Store class is introduces to emulate it within AsyncLocalStorage.

* Rename `implementation` to `driver`.

Separating name refactor out of previous commit for readability

* Allow users to specify StorageDriver

* Add env to determine StorageDriver in test

* Default to CLS_HOOKED when no option specified

* Uninstall cross-env

* Typo in getBestSupportedDriverConstructor

* Remove TODO from error message

* Update comment

* fix: remove unsupported node: protocol

* cicd: add matrix strategy to ci pipeline

* docs: describe StorageDriver enum in the readme

---------

Co-authored-by: aliheym <aliheym.dev@gmail.com>